Promoted fs_test.HttpMessage to its own httptest package
so it can be used by `fs` and `teapot`
