+++ fs/directory_listing_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/kkrull/gohttp/fs"
 	"github.com/kkrull/gohttp/http"
+	"github.com/kkrull/gohttp/httptest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -13,7 +14,7 @@ var _ = Describe("DirectoryListing", func() {
 	Describe("#WriteTo", func() {
 		var (
 			listing http.Response
-			message *HttpMessage
+			message *httptest.HttpMessage
 			err     error
 		)
 
@@ -22,7 +23,7 @@ var _ = Describe("DirectoryListing", func() {
 				output := &bytes.Buffer{}
 				listing = &fs.DirectoryListing{Files: []string{}}
 				listing.WriteTo(output)
-				message = &HttpMessage{Text: output.String()}
+				message = &httptest.HttpMessage{Text: output.String()}
 			})
 
 			It("returns no error", func() {
@@ -46,7 +47,7 @@ var _ = Describe("DirectoryListing", func() {
 				output := &bytes.Buffer{}
 				listing = &fs.DirectoryListing{Files: []string{}}
 				listing.WriteTo(output)
-				message = &HttpMessage{Text: output.String()}
+				message = &httptest.HttpMessage{Text: output.String()}
 			})
 
 			It("has an empty list of file names", func() {
@@ -61,7 +62,7 @@ var _ = Describe("DirectoryListing", func() {
 					Files:      []string{"one", "two"},
 					HrefPrefix: "/files"}
 				listing.WriteTo(output)
-				message = &HttpMessage{Text: output.String()}
+				message = &httptest.HttpMessage{Text: output.String()}
 			})
 
 			It("lists links to the files, using absolute paths with the given prefix", func() {
