+++ fs/requests_test.go
@@ -7,6 +7,7 @@ import (
 	"path"
 
 	"github.com/kkrull/gohttp/fs"
+	"github.com/kkrull/gohttp/httptest"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -16,7 +17,7 @@ var _ = Describe("Controller", func() {
 		controller *fs.Controller
 		basePath   string
 
-		response       *HttpMessage
+		response       *httptest.HttpMessage
 		responseBuffer *bytes.Buffer
 	)
 
@@ -30,7 +31,7 @@ var _ = Describe("Controller", func() {
 		Context("when the resolved target does not exist", func() {
 			BeforeEach(func() {
 				controller.Get(responseBuffer, "/missing.txt")
-				response = &HttpMessage{Text: responseBuffer.String()}
+				response = &httptest.HttpMessage{Text: responseBuffer.String()}
 			})
 
 			It("Responds with 404 Not Found", func() {
@@ -52,7 +53,7 @@ var _ = Describe("Controller", func() {
 				existingFile := path.Join(basePath, "readable.txt")
 				Expect(createTextFile(existingFile, "A")).To(Succeed())
 				controller.Get(responseBuffer, "/readable.txt")
-				response = &HttpMessage{Text: responseBuffer.String()}
+				response = &httptest.HttpMessage{Text: responseBuffer.String()}
 			})
 
 			It("responds with 200 OK", func() {
@@ -77,7 +78,7 @@ var _ = Describe("Controller", func() {
 
 			It("sets Content-Type to the MIME type registered for that extension", func() {
 				controller.Get(responseBuffer, "/image.jpeg")
-				response = &HttpMessage{Text: responseBuffer.String()}
+				response = &httptest.HttpMessage{Text: responseBuffer.String()}
 				response.HeaderShould("Content-Type", Equal("image/jpeg"))
 			})
 		})
@@ -90,7 +91,7 @@ var _ = Describe("Controller", func() {
 
 			It("sets Content-Type to text/plain", func() {
 				controller.Get(responseBuffer, "/assumed-text")
-				response = &HttpMessage{Text: responseBuffer.String()}
+				response = &httptest.HttpMessage{Text: responseBuffer.String()}
 				response.HeaderShould("Content-Type", Equal("text/plain"))
 			})
 		})
@@ -103,7 +104,7 @@ var _ = Describe("Controller", func() {
 
 			It("responds with 200 OK", func() {
 				controller.Get(responseBuffer, "/")
-				response = &HttpMessage{Text: responseBuffer.String()}
+				response = &httptest.HttpMessage{Text: responseBuffer.String()}
 				response.StatusShouldBe(200, "OK")
 			})
 		})
@@ -112,15 +113,15 @@ var _ = Describe("Controller", func() {
 	Describe("#Head", func() {
 		var (
 			getResponseBuffer = &bytes.Buffer{}
-			getResponse       *HttpMessage
+			getResponse       *httptest.HttpMessage
 		)
 
 		BeforeEach(func() {
 			controller.Get(getResponseBuffer, "/missing.txt")
-			getResponse = &HttpMessage{Text: getResponseBuffer.String()}
+			getResponse = &httptest.HttpMessage{Text: getResponseBuffer.String()}
 
 			controller.Head(responseBuffer, "/missing.txt")
-			response = &HttpMessage{Text: responseBuffer.String()}
+			response = &httptest.HttpMessage{Text: responseBuffer.String()}
 		})
 
 		It("responds with the same status as #Get would have", func() {
